+++ geoparser/__init__.py
@@ -1,2 +1,3 @@
+from geoparser.annotator import GeoparserAnnotator
 from geoparser.geoparser import Geoparser
 from geoparser.trainer import GeoparserTrainer
