+++ geoparser/annotator/__init__.py
@@ -1,0 +1,1 @@
+from .annotator import GeoparserAnnotator
