+++ geoparser/annotator/templates/start_new_session.html
