+++ geoparser/gazetteers/__init__.py
@@ -1,0 +1,2 @@
+from .geonames import GeoNames
+from .swissnames3d import SwissNames3D
