+++ geoparser/geodoc/__init__.py
@@ -1,0 +1,1 @@
+from .geodoc import GeoDoc
