+++ geoparser/geospan/__init__.py
@@ -1,0 +1,1 @@
+from .geospan import GeoSpan
