+++ geoparser/tests/gazetteer/test_gazetteer.py
@@ -8,8 +8,8 @@ import pytest
 from requests_mock.mocker import Mocker
 
 from geoparser.config.models import Column, GazetteerData
-from geoparser.gazetteer import LocalDBGazetteer
-from geoparser.geonames import GeoNames
+from geoparser.gazetteers import GeoNames
+from geoparser.gazetteers.gazetteer import LocalDBGazetteer
 from geoparser.tests.utils import execute_query, get_static_test_file, make_concrete
