+++ geoparser/tests/gazetteer/test_swissnames3d.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 
-from geoparser.swissnames3d import SwissNames3D
+from geoparser.gazetteers import SwissNames3D
 from geoparser.tests.utils import execute_query, get_static_test_file
