+++ geoparser/tests/trainer/test_trainer.py
